requirements.txt
paho-mqtt>=1.6.1,<2
influxdb-client>=1.38.0
